Wrap WgetDownload in LimitConcurrent 4
From the IRC discussions, 4 appears to be the recommended concurrency level for
this project

With this change, you can run run-pipeline with --concurrent 20 and still only
have up to 4 downloads going at the same time. The reason is to allow more
finished jobs to wait in line for RsyncUpload and improve bandwidth utilization.
This, of course, increases peak disk usage, so users with low disk space should
still not raise the concurrent level.

Suppose we are running with --concurrent 4. Sometimes we can have a big job in
RsyncUpload and 3 0MB jobs waiting for RsyncUpload. The upload bandwidth is
saturated and the download bandwidth sits idle. Then, when the big job finishes
uploading, all the 0MB jobs finish uploading instantly. Now all 4 threads are
downloading and the upload bandwidth sits idle. If we allow more jobs to wait
for RsyncUpload by wrapping WgetDownload in LimitConcurrent and running with
--concurrent 20, there is a higher chance that one of the jobs waiting for
upload will not be 0MB, and the upload bandwidth will continue to be utilized
when the big job finishes uploading.